MrX App Builder /COLAB/requirements.txt
# Core Dependencies
tree-sitter>=0.21.0
tree-sitter-languages>=1.10.2

# Embedding Models
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector Search
faiss-cpu>=1.7.4

# LLM Clients
openai>=1.0.0
requests>=2.31.0
anthropic>=0.25.0

# Git Operations
GitPython>=3.1.40

# Google Services
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
diff-match-patch>=20200713
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0